quell warnings for virtual destructors, unused vars, and size_t types
svn:revision:34226
svn:branch:trunk
svn:account:brlcad

+++ include/opennurbs_ext.h
@@ -162,7 +162,7 @@ namespace brlcad {
     template<class BV>
     inline
     BVNode<BV>::BVNode(const BV& node) : m_node(node) {
-	for (int i = 0; i < 3; i++) {
+	for (size_t i = 0; i < 3; i++) {
 	    double d = m_node.m_max[i] - m_node.m_min[i];
 	    if (ON_NearZero(d, ON_ZERO_TOLERANCE)) {
 		m_node.m_min[i] -= 0.001;
@@ -175,7 +175,7 @@ namespace brlcad {
     inline
     BVNode<BV>::~BVNode() {
 	// delete the children
-	for (int i = 0; i < m_children.size(); i++) {
+	for (size_t i = 0; i < m_children.size(); i++) {
 	    delete m_children[i];
 	}
     }
@@ -225,7 +225,7 @@ namespace brlcad {
     BVNode<BV>::intersectedBy(ON_Ray& ray, double* tnear_opt, double* tfar_opt) {
 	double tnear = -real.infinity();
 	double tfar = real.infinity();
-	for (int i = 0; i < 3; i++) {
+	for (size_t i = 0; i < 3; i++) {
 	    if (ON_NearZero(ray.m_dir[i])) {
 		if (ray.m_origin[i] < m_node.m_min[i] || ray.m_origin[i] > m_node.m_max[i])
 		    return false;
@@ -256,7 +256,7 @@ namespace brlcad {
 //       for (std::list<BVNode<BV>*>::iterator j = m_children.begin(); j != m_children.end(); j++) {
 // 	(*j)->intersectsHierarchy(ray, results_opt);
 //       }
-	    for (int i = 0; i < m_children.size(); i++) {
+	    for (size_t i = 0; i < m_children.size(); i++) {
 		m_children[i]->intersectsHierarchy(ray, results_opt);
 	    }
 	}
@@ -268,7 +268,7 @@ namespace brlcad {
     int
     BVNode<BV>::depth() {
 	int d = 0;
-	for (int i = 0; i < m_children.size(); i++) {
+	for (size_t i = 0; i < m_children.size(); i++) {
 	    d = 1 + max(d, m_children[i]->depth());
 	}
 	return d;
@@ -278,7 +278,7 @@ namespace brlcad {
     void
     BVNode<BV>::getLeaves(list<BVNode<BV>*>& out_leaves) {
 	if (m_children.size() > 0) {
-	    for (int i = 0; i < m_children.size(); i++) {
+	    for (size_t i = 0; i < m_children.size(); i++) {
 		m_children[i]->getLeaves(out_leaves);
 	    }
 	} else {
@@ -310,7 +310,7 @@ namespace brlcad {
 	TRACE("getClosestPointEstimate(" << PT(pt) << ")");
 	if (m_children.size() > 0) {
 	    BBNode* closestNode = m_children[0];
-	    for (int i = 1; i < m_children.size(); i++) {
+	    for (size_t i = 1; i < m_children.size(); i++) {
 		closestNode = closer(pt, closestNode, m_children[i]);
 		TRACE("\t" << PT(closestNode->m_estimate));
 	    }
@@ -380,7 +380,7 @@ namespace brlcad {
 	int mini = 0;
 	double mindist = pt.DistanceTo(corners[mini]);
 	double tmpdist;
-	for (int i = 1; i < 5; i++) {
+	for (size_t i = 1; i < 5; i++) {
 	    tmpdist = pt.DistanceTo(corners[i]);
 	    TRACE("\t" << mindist << " < " << tmpdist);
 	    if (tmpdist < mindist) {
